k8s/watchers: refactor watchers to dedicated files
