+++ pkg/k8s/watchers/watcher.go
@@ -16,7 +16,6 @@ package watchers
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"net"
 	"net/url"
@@ -24,23 +23,10 @@ import (
 	"sync"
 	"time"
 
-	"github.com/cilium/cilium/pkg/annotation"
-	"github.com/cilium/cilium/pkg/comparator"
 	"github.com/cilium/cilium/pkg/controller"
 	"github.com/cilium/cilium/pkg/endpoint"
-	"github.com/cilium/cilium/pkg/endpoint/regeneration"
-	"github.com/cilium/cilium/pkg/identity"
-	"github.com/cilium/cilium/pkg/ipcache"
 	"github.com/cilium/cilium/pkg/k8s"
-	ciliumio "github.com/cilium/cilium/pkg/k8s/apis/cilium.io"
-	cilium_v2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
-	clientset "github.com/cilium/cilium/pkg/k8s/client/clientset/versioned"
-	"github.com/cilium/cilium/pkg/k8s/informer"
 	k8smetrics "github.com/cilium/cilium/pkg/k8s/metrics"
-	"github.com/cilium/cilium/pkg/k8s/types"
-	k8sUtils "github.com/cilium/cilium/pkg/k8s/utils"
-	k8sversion "github.com/cilium/cilium/pkg/k8s/version"
-	"github.com/cilium/cilium/pkg/kvstore"
 	"github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/loadbalancer"
 	"github.com/cilium/cilium/pkg/lock"
@@ -52,15 +38,9 @@ import (
 	"github.com/cilium/cilium/pkg/policy"
 	"github.com/cilium/cilium/pkg/policy/api"
 	"github.com/cilium/cilium/pkg/serializer"
-	"github.com/cilium/cilium/pkg/source"
-	"github.com/cilium/cilium/pkg/spanstat"
 
 	"github.com/sirupsen/logrus"
-	v1 "k8s.io/api/core/v1"
-	networkingv1 "k8s.io/api/networking/v1"
-	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/util/runtime"
-	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/tools/cache"
 	k8s_metrics "k8s.io/client-go/tools/metrics"
 )
@@ -90,6 +70,16 @@ const (
 	metricDelete         = "delete"
 	metricUpdate         = "update"
 )
+
+func init() {
+	// Replace error handler with our own
+	runtime.ErrorHandlers = []func(error){
+		k8s.K8sErrorHandler,
+	}
+
+	k8sMetric := &k8sMetrics{}
+	k8s_metrics.Register(k8sMetric, k8sMetric)
+}
 
 var (
 	log = logging.DefaultLogger.WithField(logfields.LogSubsys, "k8s-watcher")
@@ -179,58 +169,6 @@ func NewK8sWatcher(
 	}
 }
 
-// ruleImportMetadataCache maps the unique identifier of a CiliumNetworkPolicy
-// (namespace and name) to metadata about the importing of the rule into the
-// agent's policy repository at the time said rule was imported (revision
-// number, and if any error occurred while importing).
-type ruleImportMetadataCache struct {
-	mutex                 lock.RWMutex
-	ruleImportMetadataMap map[string]policyImportMetadata
-}
-
-type policyImportMetadata struct {
-	revision          uint64
-	policyImportError error
-}
-
-func (r *ruleImportMetadataCache) upsert(cnp *types.SlimCNP, revision uint64, importErr error) {
-	if cnp == nil {
-		return
-	}
-
-	meta := policyImportMetadata{
-		revision:          revision,
-		policyImportError: importErr,
-	}
-	podNSName := k8sUtils.GetObjNamespaceName(&cnp.ObjectMeta)
-
-	r.mutex.Lock()
-	r.ruleImportMetadataMap[podNSName] = meta
-	r.mutex.Unlock()
-}
-
-func (r *ruleImportMetadataCache) delete(cnp *types.SlimCNP) {
-	if cnp == nil {
-		return
-	}
-	podNSName := k8sUtils.GetObjNamespaceName(&cnp.ObjectMeta)
-
-	r.mutex.Lock()
-	delete(r.ruleImportMetadataMap, podNSName)
-	r.mutex.Unlock()
-}
-
-func (r *ruleImportMetadataCache) get(cnp *types.SlimCNP) (policyImportMetadata, bool) {
-	if cnp == nil {
-		return policyImportMetadata{}, false
-	}
-	podNSName := k8sUtils.GetObjNamespaceName(&cnp.ObjectMeta)
-	r.mutex.RLock()
-	policyImportMeta, ok := r.ruleImportMetadataMap[podNSName]
-	r.mutex.RUnlock()
-	return policyImportMeta, ok
-}
-
 // endpointImportMetadataCache maps the unique identifier of a Kubernetes
 // Endpoint (namespace and name) to metadata about whether translation of the
 // rules involving services that the endpoint corresponds to into the
@@ -312,16 +250,6 @@ func (*k8sMetrics) Observe(verb string, u url.URL, latency time.Duration) {
 func (*k8sMetrics) Increment(code string, method string, host string) {
 	metrics.KubernetesAPICalls.WithLabelValues(host, method, code).Inc()
 	k8smetrics.LastInteraction.Reset()
-}
-
-func init() {
-	// Replace error handler with our own
-	runtime.ErrorHandlers = []func(error){
-		k8s.K8sErrorHandler,
-	}
-
-	k8sMetric := &k8sMetrics{}
-	k8s_metrics.Register(k8sMetric, k8sMetric)
 }
 
 func (k *K8sWatcher) GetAPIGroups() []string {
@@ -471,732 +399,28 @@ func (k *K8sWatcher) EnableK8sWatcher(queueSize uint) error {
 
 	serNamespaces := serializer.NewFunctionQueue(queueSize)
 
-	_, policyController := informer.NewInformer(
-		cache.NewListWatchFromClient(k8s.Client().NetworkingV1().RESTClient(),
-			"networkpolicies", v1.NamespaceAll, fields.Everything()),
-		&networkingv1.NetworkPolicy{},
-		0,
-		cache.ResourceEventHandlerFuncs{
-			AddFunc: func(obj interface{}) {
-				var valid, equal bool
-				defer func() { k.K8sEventReceived(metricKNP, metricCreate, valid, equal) }()
-				if k8sNP := k8s.CopyObjToV1NetworkPolicy(obj); k8sNP != nil {
-					valid = true
-					swgKNP.Add()
-					serKNPs.Enqueue(func() error {
-						defer swgKNP.Done()
-						err := k.addK8sNetworkPolicyV1(k8sNP)
-						k.K8sEventProcessed(metricKNP, metricCreate, err == nil)
-						return nil
-					}, serializer.NoRetry)
-				}
-			},
-			UpdateFunc: func(oldObj, newObj interface{}) {
-				var valid, equal bool
-				defer func() { k.K8sEventReceived(metricKNP, metricUpdate, valid, equal) }()
-				if oldK8sNP := k8s.CopyObjToV1NetworkPolicy(oldObj); oldK8sNP != nil {
-					valid = true
-					if newK8sNP := k8s.CopyObjToV1NetworkPolicy(newObj); newK8sNP != nil {
-						if k8s.EqualV1NetworkPolicy(oldK8sNP, newK8sNP) {
-							equal = true
-							return
-						}
+	k.networkPoliciesInit(k8s.Client(), serKNPs, swgKNP)
 
-						swgKNP.Add()
-						serKNPs.Enqueue(func() error {
-							defer swgKNP.Done()
-							err := k.updateK8sNetworkPolicyV1(oldK8sNP, newK8sNP)
-							k.K8sEventProcessed(metricKNP, metricUpdate, err == nil)
-							return nil
-						}, serializer.NoRetry)
-					}
-				}
-			},
-			DeleteFunc: func(obj interface{}) {
-				var valid, equal bool
-				defer func() { k.K8sEventReceived(metricKNP, metricDelete, valid, equal) }()
-				k8sNP := k8s.CopyObjToV1NetworkPolicy(obj)
-				if k8sNP == nil {
-					deletedObj, ok := obj.(cache.DeletedFinalStateUnknown)
-					if !ok {
-						return
-					}
-					// Delete was not observed by the watcher but is
-					// removed from kube-apiserver. This is the last
-					// known state and the object no longer exists.
-					k8sNP = k8s.CopyObjToV1NetworkPolicy(deletedObj.Obj)
-					if k8sNP == nil {
-						return
-					}
-				}
+	k.servicesInit(k8s.Client(), serSvcs, swgSvcs)
 
-				valid = true
-				swgKNP.Add()
-				serKNPs.Enqueue(func() error {
-					defer swgKNP.Done()
-					err := k.deleteK8sNetworkPolicyV1(k8sNP)
-					k.K8sEventProcessed(metricKNP, metricDelete, err == nil)
-					return nil
-				}, serializer.NoRetry)
-			},
-		},
-		k8s.ConvertToNetworkPolicy,
-	)
-	k.blockWaitGroupToSyncResources(wait.NeverStop, swgKNP, policyController, k8sAPIGroupNetworkingV1Core)
-	go policyController.Run(wait.NeverStop)
+	k.endpointsInit(k8s.Client(), serEps, swgEps)
 
-	k.k8sAPIGroups.addAPI(k8sAPIGroupNetworkingV1Core)
+	k.ciliumNetworkPoliciesInit(ciliumNPClient, serCNPs, swgCNPs)
 
-	_, svcController := informer.NewInformer(
-		cache.NewListWatchFromClient(k8s.Client().CoreV1().RESTClient(),
-			"services", v1.NamespaceAll, fields.Everything()),
-		&v1.Service{},
-		0,
-		cache.ResourceEventHandlerFuncs{
-			AddFunc: func(obj interface{}) {
-				var valid, equal bool
-				defer func() { k.K8sEventReceived(metricService, metricCreate, valid, equal) }()
-				if k8sSvc := k8s.CopyObjToV1Services(obj); k8sSvc != nil {
-					valid = true
-					swgSvcs.Add()
-					serSvcs.Enqueue(func() error {
-						defer swgSvcs.Done()
-						err := k.addK8sServiceV1(k8sSvc, swgSvcs)
-						k.K8sEventProcessed(metricService, metricCreate, err == nil)
-						return nil
-					}, serializer.NoRetry)
-				}
-			},
-			UpdateFunc: func(oldObj, newObj interface{}) {
-				var valid, equal bool
-				defer func() { k.K8sEventReceived(metricService, metricUpdate, valid, equal) }()
-				if oldk8sSvc := k8s.CopyObjToV1Services(oldObj); oldk8sSvc != nil {
-					valid = true
-					if newk8sSvc := k8s.CopyObjToV1Services(newObj); newk8sSvc != nil {
-						if k8s.EqualV1Services(oldk8sSvc, newk8sSvc) {
-							equal = true
-							return
-						}
-
-						swgSvcs.Add()
-						serSvcs.Enqueue(func() error {
-							defer swgSvcs.Done()
-							err := k.updateK8sServiceV1(oldk8sSvc, newk8sSvc, swgSvcs)
-							k.K8sEventProcessed(metricService, metricUpdate, err == nil)
-							return nil
-						}, serializer.NoRetry)
-					}
-				}
-			},
-			DeleteFunc: func(obj interface{}) {
-				var valid, equal bool
-				defer func() { k.K8sEventReceived(metricService, metricDelete, valid, equal) }()
-				k8sSvc := k8s.CopyObjToV1Services(obj)
-				if k8sSvc == nil {
-					deletedObj, ok := obj.(cache.DeletedFinalStateUnknown)
-					if !ok {
-						return
-					}
-					// Delete was not observed by the watcher but is
-					// removed from kube-apiserver. This is the last
-					// known state and the object no longer exists.
-					k8sSvc = k8s.CopyObjToV1Services(deletedObj.Obj)
-					if k8sSvc == nil {
-						return
-					}
-				}
-
-				valid = true
-				swgSvcs.Add()
-				serSvcs.Enqueue(func() error {
-					defer swgSvcs.Done()
-					err := k.deleteK8sServiceV1(k8sSvc, swgSvcs)
-					k.K8sEventProcessed(metricService, metricDelete, err == nil)
-					return nil
-				}, serializer.NoRetry)
-			},
-		},
-		k8s.ConvertToK8sService,
-	)
-	k.blockWaitGroupToSyncResources(wait.NeverStop, swgSvcs, svcController, K8sAPIGroupServiceV1Core)
-	go svcController.Run(wait.NeverStop)
-	k.k8sAPIGroups.addAPI(K8sAPIGroupServiceV1Core)
-
-	_, endpointController := informer.NewInformer(
-		cache.NewListWatchFromClient(k8s.Client().CoreV1().RESTClient(),
-			"endpoints", v1.NamespaceAll,
-			fields.ParseSelectorOrDie(option.Config.K8sWatcherEndpointSelector),
-		),
-		&v1.Endpoints{},
-		0,
-		cache.ResourceEventHandlerFuncs{
-			AddFunc: func(obj interface{}) {
-				var valid, equal bool
-				defer func() { k.K8sEventReceived(metricEndpoint, metricCreate, valid, equal) }()
-				if k8sEP := k8s.CopyObjToV1Endpoints(obj); k8sEP != nil {
-					valid = true
-					swgEps.Add()
-					serEps.Enqueue(func() error {
-						defer swgEps.Done()
-						err := k.addK8sEndpointV1(k8sEP, swgEps)
-						k.K8sEventProcessed(metricEndpoint, metricCreate, err == nil)
-						return nil
-					}, serializer.NoRetry)
-				}
-			},
-			UpdateFunc: func(oldObj, newObj interface{}) {
-				var valid, equal bool
-				defer func() { k.K8sEventReceived(metricEndpoint, metricUpdate, valid, equal) }()
-				if oldk8sEP := k8s.CopyObjToV1Endpoints(oldObj); oldk8sEP != nil {
-					valid = true
-					if newk8sEP := k8s.CopyObjToV1Endpoints(newObj); newk8sEP != nil {
-						if k8s.EqualV1Endpoints(oldk8sEP, newk8sEP) {
-							equal = true
-							return
-						}
-
-						swgEps.Add()
-						serEps.Enqueue(func() error {
-							defer swgEps.Done()
-							err := k.updateK8sEndpointV1(oldk8sEP, newk8sEP, swgEps)
-							k.K8sEventProcessed(metricEndpoint, metricUpdate, err == nil)
-							return nil
-						}, serializer.NoRetry)
-					}
-				}
-			},
-			DeleteFunc: func(obj interface{}) {
-				var valid, equal bool
-				defer func() { k.K8sEventReceived(metricEndpoint, metricDelete, valid, equal) }()
-				k8sEP := k8s.CopyObjToV1Endpoints(obj)
-				if k8sEP == nil {
-					deletedObj, ok := obj.(cache.DeletedFinalStateUnknown)
-					if !ok {
-						return
-					}
-					// Delete was not observed by the watcher but is
-					// removed from kube-apiserver. This is the last
-					// known state and the object no longer exists.
-					k8sEP = k8s.CopyObjToV1Endpoints(deletedObj.Obj)
-					if k8sEP == nil {
-						return
-					}
-				}
-				valid = true
-				swgEps.Add()
-				serEps.Enqueue(func() error {
-					defer swgEps.Done()
-					err := k.deleteK8sEndpointV1(k8sEP, swgEps)
-					k.K8sEventProcessed(metricEndpoint, metricDelete, err == nil)
-					return nil
-				}, serializer.NoRetry)
-			},
-		},
-		k8s.ConvertToK8sEndpoints,
-	)
-	k.blockWaitGroupToSyncResources(wait.NeverStop, swgEps, endpointController, K8sAPIGroupEndpointV1Core)
-	go endpointController.Run(wait.NeverStop)
-	k.k8sAPIGroups.addAPI(K8sAPIGroupEndpointV1Core)
-
-	var (
-		cnpEventStore    cache.Store
-		cnpConverterFunc informer.ConvertFunc
-	)
-	cnpStore := cache.NewStore(cache.DeletionHandlingMetaNamespaceKeyFunc)
-	switch {
-	case k8sversion.Capabilities().Patch:
-		// k8s >= 1.13 does not require a store to update CNP status so
-		// we don't even need to keep the status of a CNP with us.
-		cnpConverterFunc = k8s.ConvertToCNP
-	default:
-		cnpEventStore = cnpStore
-		cnpConverterFunc = k8s.ConvertToCNPWithStatus
-	}
-
-	ciliumV2Controller := informer.NewInformerWithStore(
-		cache.NewListWatchFromClient(ciliumNPClient.CiliumV2().RESTClient(),
-			"ciliumnetworkpolicies", v1.NamespaceAll, fields.Everything()),
-		&cilium_v2.CiliumNetworkPolicy{},
-		0,
-		cache.ResourceEventHandlerFuncs{
-			AddFunc: func(obj interface{}) {
-				var valid, equal bool
-				defer func() { k.K8sEventReceived(metricCNP, metricCreate, valid, equal) }()
-				if cnp := k8s.CopyObjToV2CNP(obj); cnp != nil {
-					valid = true
-					swgCNPs.Add()
-					serCNPs.Enqueue(func() error {
-						defer swgCNPs.Done()
-						if cnp.RequiresDerivative() {
-							return nil
-						}
-						err := k.addCiliumNetworkPolicyV2(ciliumNPClient, cnpEventStore, cnp)
-						k.K8sEventProcessed(metricCNP, metricCreate, err == nil)
-						return nil
-					}, serializer.NoRetry)
-				}
-			},
-			UpdateFunc: func(oldObj, newObj interface{}) {
-				var valid, equal bool
-				defer func() { k.K8sEventReceived(metricCNP, metricUpdate, valid, equal) }()
-				if oldCNP := k8s.CopyObjToV2CNP(oldObj); oldCNP != nil {
-					valid = true
-					if newCNP := k8s.CopyObjToV2CNP(newObj); newCNP != nil {
-						if k8s.EqualV2CNP(oldCNP, newCNP) {
-							equal = true
-							return
-						}
-
-						swgCNPs.Add()
-						serCNPs.Enqueue(func() error {
-							defer swgCNPs.Done()
-							if newCNP.RequiresDerivative() {
-								return nil
-							}
-
-							err := k.updateCiliumNetworkPolicyV2(ciliumNPClient, cnpEventStore, oldCNP, newCNP)
-							k.K8sEventProcessed(metricCNP, metricUpdate, err == nil)
-							return nil
-						}, serializer.NoRetry)
-					}
-				}
-			},
-			DeleteFunc: func(obj interface{}) {
-				var valid, equal bool
-				defer func() { k.K8sEventReceived(metricCNP, metricDelete, valid, equal) }()
-				cnp := k8s.CopyObjToV2CNP(obj)
-				if cnp == nil {
-					deletedObj, ok := obj.(cache.DeletedFinalStateUnknown)
-					if !ok {
-						return
-					}
-					// Delete was not observed by the watcher but is
-					// removed from kube-apiserver. This is the last
-					// known state and the object no longer exists.
-					cnp = k8s.CopyObjToV2CNP(deletedObj.Obj)
-					if cnp == nil {
-						return
-					}
-				}
-				valid = true
-				swgCNPs.Add()
-				serCNPs.Enqueue(func() error {
-					defer swgCNPs.Done()
-					err := k.deleteCiliumNetworkPolicyV2(cnp)
-					k.K8sEventProcessed(metricCNP, metricDelete, err == nil)
-					return nil
-				}, serializer.NoRetry)
-			},
-		},
-		cnpConverterFunc,
-		cnpStore,
-	)
-	k.blockWaitGroupToSyncResources(wait.NeverStop, swgCNPs, ciliumV2Controller, k8sAPIGroupCiliumNetworkPolicyV2)
-	go ciliumV2Controller.Run(wait.NeverStop)
-	k.k8sAPIGroups.addAPI(k8sAPIGroupCiliumNetworkPolicyV2)
-
-	asyncControllers := sync.WaitGroup{}
+	asyncControllers := &sync.WaitGroup{}
 	asyncControllers.Add(1)
-
-	// CiliumNode objects are used for node discovery until the key-value
-	// store is connected
-	go func() {
-		var once sync.Once
-		for {
-			swgNodes := lock.NewStoppableWaitGroup()
-			_, ciliumNodeInformer := informer.NewInformer(
-				cache.NewListWatchFromClient(ciliumNPClient.CiliumV2().RESTClient(),
-					"ciliumnodes", v1.NamespaceAll, fields.Everything()),
-				&cilium_v2.CiliumNode{},
-				0,
-				cache.ResourceEventHandlerFuncs{
-					AddFunc: func(obj interface{}) {
-						var valid, equal bool
-						defer func() { k.K8sEventReceived(metricCiliumNode, metricCreate, valid, equal) }()
-						if ciliumNode, ok := obj.(*cilium_v2.CiliumNode); ok {
-							valid = true
-							n := node.ParseCiliumNode(ciliumNode)
-							if n.IsLocal() {
-								return
-							}
-							swgNodes.Add()
-							serNodes.Enqueue(func() error {
-								defer swgNodes.Done()
-								k.nodeDiscoverManager.NodeUpdated(n)
-								k.K8sEventProcessed(metricCiliumNode, metricCreate, true)
-								return nil
-							}, serializer.NoRetry)
-						}
-					},
-					UpdateFunc: func(oldObj, newObj interface{}) {
-						var valid, equal bool
-						defer func() { k.K8sEventReceived(metricCiliumNode, metricUpdate, valid, equal) }()
-						if ciliumNode, ok := newObj.(*cilium_v2.CiliumNode); ok {
-							valid = true
-							n := node.ParseCiliumNode(ciliumNode)
-							if n.IsLocal() {
-								return
-							}
-							swgNodes.Add()
-							serNodes.Enqueue(func() error {
-								defer swgNodes.Done()
-								k.nodeDiscoverManager.NodeUpdated(n)
-								k.K8sEventProcessed(metricCiliumNode, metricUpdate, true)
-								return nil
-							}, serializer.NoRetry)
-						}
-					},
-					DeleteFunc: func(obj interface{}) {
-						var valid, equal bool
-						defer func() { k.K8sEventReceived(metricCiliumNode, metricDelete, valid, equal) }()
-						ciliumNode := k8s.CopyObjToCiliumNode(obj)
-						if ciliumNode == nil {
-							deletedObj, ok := obj.(cache.DeletedFinalStateUnknown)
-							if !ok {
-								return
-							}
-							// Delete was not observed by the watcher but is
-							// removed from kube-apiserver. This is the last
-							// known state and the object no longer exists.
-							ciliumNode = k8s.CopyObjToCiliumNode(deletedObj.Obj)
-							if ciliumNode == nil {
-								return
-							}
-						}
-						valid = true
-						n := node.ParseCiliumNode(ciliumNode)
-						swgNodes.Add()
-						serNodes.Enqueue(func() error {
-							defer swgNodes.Done()
-							k.nodeDiscoverManager.NodeDeleted(n)
-							return nil
-						}, serializer.NoRetry)
-					},
-				},
-				k8s.ConvertToCiliumNode,
-			)
-			isConnected := make(chan struct{})
-			// once isConnected is closed, it will stop waiting on caches to be
-			// synchronized.
-			k.blockWaitGroupToSyncResources(isConnected, swgNodes, ciliumNodeInformer, k8sAPIGroupCiliumNodeV2)
-
-			once.Do(func() {
-				// Signalize that we have put node controller in the wait group
-				// to sync resources.
-				asyncControllers.Done()
-			})
-			k.k8sAPIGroups.addAPI(k8sAPIGroupCiliumNodeV2)
-			go ciliumNodeInformer.Run(isConnected)
-
-			<-kvstore.Client().Connected()
-			close(isConnected)
-
-			log.Info("Connected to key-value store, stopping CiliumNode watcher")
-
-			k.k8sAPIGroups.removeAPI(k8sAPIGroupCiliumNodeV2)
-			// Create a new node controller when we are disconnected with the
-			// kvstore
-			<-kvstore.Client().Disconnected()
-
-			log.Info("Disconnected from key-value store, restarting CiliumNode watcher")
-		}
-	}()
+	go k.ciliumNodeInit(ciliumNPClient, serNodes, asyncControllers)
 
 	asyncControllers.Add(1)
-
-	// CiliumEndpoint objects are used for ipcache discovery until the
-	// key-value store is connected
-	go func() {
-		var once sync.Once
-		for {
-			swgCiliumEndpoints := lock.NewStoppableWaitGroup()
-			_, ciliumEndpointInformer := informer.NewInformer(
-				cache.NewListWatchFromClient(ciliumNPClient.CiliumV2().RESTClient(),
-					"ciliumendpoints", v1.NamespaceAll, fields.Everything()),
-				&cilium_v2.CiliumEndpoint{},
-				0,
-				cache.ResourceEventHandlerFuncs{
-					AddFunc: func(obj interface{}) {
-						var valid, equal bool
-						defer func() { k.K8sEventReceived(metricCiliumEndpoint, metricCreate, valid, equal) }()
-						if ciliumEndpoint, ok := obj.(*types.CiliumEndpoint); ok {
-							valid = true
-							endpoint := ciliumEndpoint.DeepCopy()
-							swgCiliumEndpoints.Add()
-							serCiliumEndpoints.Enqueue(func() error {
-								defer swgCiliumEndpoints.Done()
-								endpointUpdated(endpoint)
-								k.K8sEventProcessed(metricCiliumEndpoint, metricCreate, true)
-								return nil
-							}, serializer.NoRetry)
-						}
-					},
-					UpdateFunc: func(oldObj, newObj interface{}) {
-						var valid, equal bool
-						defer func() { k.K8sEventReceived(metricCiliumEndpoint, metricUpdate, valid, equal) }()
-						if ciliumEndpoint, ok := newObj.(*types.CiliumEndpoint); ok {
-							valid = true
-							endpoint := ciliumEndpoint.DeepCopy()
-							swgCiliumEndpoints.Add()
-							serCiliumEndpoints.Enqueue(func() error {
-								defer swgCiliumEndpoints.Done()
-								endpointUpdated(endpoint)
-								k.K8sEventProcessed(metricCiliumEndpoint, metricUpdate, true)
-								return nil
-							}, serializer.NoRetry)
-						}
-					},
-					DeleteFunc: func(obj interface{}) {
-						var valid, equal bool
-						defer func() { k.K8sEventReceived(metricCiliumEndpoint, metricDelete, valid, equal) }()
-						ciliumEndpoint := k8s.CopyObjToCiliumEndpoint(obj)
-						if ciliumEndpoint == nil {
-							deletedObj, ok := obj.(cache.DeletedFinalStateUnknown)
-							if !ok {
-								return
-							}
-							// Delete was not observed by the watcher but is
-							// removed from kube-apiserver. This is the last
-							// known state and the object no longer exists.
-							ciliumEndpoint = k8s.CopyObjToCiliumEndpoint(deletedObj.Obj)
-							if ciliumEndpoint == nil {
-								return
-							}
-						}
-						valid = true
-						swgCiliumEndpoints.Add()
-						serCiliumEndpoints.Enqueue(func() error {
-							defer swgCiliumEndpoints.Done()
-							endpointDeleted(ciliumEndpoint)
-							return nil
-						}, serializer.NoRetry)
-					},
-				},
-				k8s.ConvertToCiliumEndpoint,
-			)
-			isConnected := make(chan struct{})
-			// once isConnected is closed, it will stop waiting on caches to be
-			// synchronized.
-			k.blockWaitGroupToSyncResources(isConnected, swgCiliumEndpoints, ciliumEndpointInformer, k8sAPIGroupCiliumEndpointV2)
-
-			once.Do(func() {
-				// Signalize that we have put node controller in the wait group
-				// to sync resources.
-				asyncControllers.Done()
-			})
-			k.k8sAPIGroups.addAPI(k8sAPIGroupCiliumEndpointV2)
-			go ciliumEndpointInformer.Run(isConnected)
-
-			<-kvstore.Client().Connected()
-			close(isConnected)
-
-			log.Info("Connected to key-value store, stopping CiliumEndpoint watcher")
-
-			k.k8sAPIGroups.removeAPI(k8sAPIGroupCiliumEndpointV2)
-			// Create a new node controller when we are disconnected with the
-			// kvstore
-			<-kvstore.Client().Disconnected()
-
-			log.Info("Disconnected from key-value store, restarting CiliumEndpoint watcher")
-		}
-	}()
+	go k.ciliumEndpointsInit(ciliumNPClient, serCiliumEndpoints, asyncControllers)
 
 	asyncControllers.Add(1)
-	go func() {
-		var once sync.Once
-		for {
-			swgPods := lock.NewStoppableWaitGroup()
-			createPodController := func(fieldSelector fields.Selector) cache.Controller {
-				_, podController := informer.NewInformer(
-					cache.NewListWatchFromClient(k8s.Client().CoreV1().RESTClient(),
-						"pods", v1.NamespaceAll, fieldSelector),
-					&v1.Pod{},
-					0,
-					cache.ResourceEventHandlerFuncs{
-						AddFunc: func(obj interface{}) {
-							var valid, equal bool
-							defer func() { k.K8sEventReceived(metricPod, metricCreate, valid, equal) }()
-							if pod := k8s.CopyObjToV1Pod(obj); pod != nil {
-								valid = true
-								swgPods.Add()
-								serPods.Enqueue(func() error {
-									defer swgPods.Done()
-									err := k.addK8sPodV1(pod)
-									k.K8sEventProcessed(metricPod, metricCreate, err == nil)
-									return nil
-								}, serializer.NoRetry)
-							}
-						},
-						UpdateFunc: func(oldObj, newObj interface{}) {
-							var valid, equal bool
-							defer func() { k.K8sEventReceived(metricPod, metricUpdate, valid, equal) }()
-							if oldPod := k8s.CopyObjToV1Pod(oldObj); oldPod != nil {
-								valid = true
-								if newPod := k8s.CopyObjToV1Pod(newObj); newPod != nil {
-									if k8s.EqualV1Pod(oldPod, newPod) {
-										equal = true
-										return
-									}
-									swgPods.Add()
-									serPods.Enqueue(func() error {
-										defer swgPods.Done()
-										err := k.updateK8sPodV1(oldPod, newPod)
-										k.K8sEventProcessed(metricPod, metricUpdate, err == nil)
-										return nil
-									}, serializer.NoRetry)
-								}
-							}
-						},
-						DeleteFunc: func(obj interface{}) {
-							var valid, equal bool
-							defer func() { k.K8sEventReceived(metricPod, metricDelete, valid, equal) }()
-							if pod := k8s.CopyObjToV1Pod(obj); pod != nil {
-								valid = true
-								swgPods.Add()
-								serPods.Enqueue(func() error {
-									defer swgPods.Done()
-									err := k.deleteK8sPodV1(pod)
-									k.K8sEventProcessed(metricPod, metricDelete, err == nil)
-									return nil
-								}, serializer.NoRetry)
-							}
-						},
-					},
-					k8s.ConvertToPod,
-				)
-				return podController
-			}
-			podController := createPodController(fields.Everything())
+	go k.podsInit(k8s.Client(), serPods, asyncControllers)
 
-			isConnected := make(chan struct{})
-			// once isConnected is closed, it will stop waiting on caches to be
-			// synchronized.
-			k.blockWaitGroupToSyncResources(isConnected, swgPods, podController, k8sAPIGroupPodV1Core)
-			once.Do(func() {
-				asyncControllers.Done()
-				k.k8sAPIGroups.addAPI(k8sAPIGroupPodV1Core)
-			})
-			go podController.Run(isConnected)
-
-			if !option.Config.K8sEventHandover {
-				return
-			}
-
-			// Replace pod controller by only receiving events from our own
-			// node once we are connected to the kvstore.
-
-			<-kvstore.Client().Connected()
-			close(isConnected)
-
-			log.WithField(logfields.Node, node.GetName()).Info("Connected to KVStore, watching for pod events on node")
-			// Only watch for pod events for our node.
-			podController = createPodController(fields.ParseSelectorOrDie("spec.nodeName=" + node.GetName()))
-			isConnected = make(chan struct{})
-			go podController.Run(isConnected)
-
-			// Create a new pod controller when we are disconnected with the
-			// kvstore
-			<-kvstore.Client().Disconnected()
-			close(isConnected)
-			log.Info("Disconnected from KVStore, watching for pod events all nodes")
-		}
-	}()
-
-	_, namespaceController := informer.NewInformer(
-		cache.NewListWatchFromClient(k8s.Client().CoreV1().RESTClient(),
-			"namespaces", v1.NamespaceAll, fields.Everything()),
-		&v1.Namespace{},
-		0,
-		cache.ResourceEventHandlerFuncs{
-			// AddFunc does not matter since the endpoint will fetch
-			// namespace labels when the endpoint is created
-			// DelFunc does not matter since, when a namespace is deleted, all
-			// pods belonging to that namespace are also deleted.
-			UpdateFunc: func(oldObj, newObj interface{}) {
-				var valid, equal bool
-				defer func() { k.K8sEventReceived(metricNS, metricUpdate, valid, equal) }()
-				if oldNS := k8s.CopyObjToV1Namespace(oldObj); oldNS != nil {
-					valid = true
-					if newNS := k8s.CopyObjToV1Namespace(newObj); newNS != nil {
-						if k8s.EqualV1Namespace(oldNS, newNS) {
-							equal = true
-							return
-						}
-
-						serNamespaces.Enqueue(func() error {
-							err := k.updateK8sV1Namespace(oldNS, newNS)
-							k.K8sEventProcessed(metricNS, metricUpdate, err == nil)
-							return nil
-						}, serializer.NoRetry)
-					}
-				}
-			},
-		},
-		k8s.ConvertToNamespace,
-	)
-
-	go namespaceController.Run(wait.NeverStop)
-	k.k8sAPIGroups.addAPI(k8sAPIGroupNamespaceV1Core)
+	go k.namespacesInit(k8s.Client(), serNamespaces)
 
 	asyncControllers.Wait()
 
-	return nil
-}
-
-func (k *K8sWatcher) addK8sNetworkPolicyV1(k8sNP *types.NetworkPolicy) error {
-	scopedLog := log.WithField(logfields.K8sAPIVersion, k8sNP.TypeMeta.APIVersion)
-	rules, err := k8s.ParseNetworkPolicy(k8sNP.NetworkPolicy)
-	if err != nil {
-		scopedLog.WithError(err).WithFields(logrus.Fields{
-			logfields.CiliumNetworkPolicy: logfields.Repr(k8sNP),
-		}).Error("Error while parsing k8s kubernetes NetworkPolicy")
-		return err
-	}
-	scopedLog = scopedLog.WithField(logfields.K8sNetworkPolicyName, k8sNP.ObjectMeta.Name)
-
-	opts := policy.AddOptions{Replace: true, Source: metrics.LabelEventSourceK8s}
-	if _, err := k.policyManager.PolicyAdd(rules, &opts); err != nil {
-		scopedLog.WithError(err).WithFields(logrus.Fields{
-			logfields.CiliumNetworkPolicy: logfields.Repr(rules),
-		}).Error("Unable to add NetworkPolicy rules to policy repository")
-		return err
-	}
-
-	scopedLog.Info("NetworkPolicy successfully added")
-	return nil
-}
-
-func (k *K8sWatcher) updateK8sNetworkPolicyV1(oldk8sNP, newk8sNP *types.NetworkPolicy) error {
-	log.WithFields(logrus.Fields{
-		logfields.K8sAPIVersion:                 oldk8sNP.TypeMeta.APIVersion,
-		logfields.K8sNetworkPolicyName + ".old": oldk8sNP.ObjectMeta.Name,
-		logfields.K8sNamespace + ".old":         oldk8sNP.ObjectMeta.Namespace,
-		logfields.K8sNetworkPolicyName:          newk8sNP.ObjectMeta.Name,
-		logfields.K8sNamespace:                  newk8sNP.ObjectMeta.Namespace,
-	}).Debug("Received policy update")
-
-	return k.addK8sNetworkPolicyV1(newk8sNP)
-}
-
-func (k *K8sWatcher) deleteK8sNetworkPolicyV1(k8sNP *types.NetworkPolicy) error {
-	labels := k8s.GetPolicyLabelsv1(k8sNP.NetworkPolicy)
-
-	if labels == nil {
-		log.Fatalf("provided v1 NetworkPolicy is nil, so cannot delete it")
-	}
-
-	scopedLog := log.WithFields(logrus.Fields{
-		logfields.K8sNetworkPolicyName: k8sNP.ObjectMeta.Name,
-		logfields.K8sNamespace:         k8sNP.ObjectMeta.Namespace,
-		logfields.K8sAPIVersion:        k8sNP.TypeMeta.APIVersion,
-		logfields.Labels:               logfields.Repr(labels),
-	})
-	if _, err := k.policyManager.PolicyDelete(labels); err != nil {
-		scopedLog.WithError(err).Error("Error while deleting k8s NetworkPolicy")
-		return err
-	}
-
-	scopedLog.Info("NetworkPolicy successfully removed")
 	return nil
 }
 
@@ -1281,35 +505,6 @@ func (k *K8sWatcher) k8sServiceHandler() {
 
 func (k *K8sWatcher) RunK8sServiceHandler() {
 	go k.k8sServiceHandler()
-}
-
-func (k *K8sWatcher) addK8sServiceV1(svc *types.Service, swg *lock.StoppableWaitGroup) error {
-	k.K8sSvcCache.UpdateService(svc, swg)
-	return nil
-}
-
-func (k *K8sWatcher) updateK8sServiceV1(oldSvc, newSvc *types.Service, swg *lock.StoppableWaitGroup) error {
-	return k.addK8sServiceV1(newSvc, swg)
-}
-
-func (k *K8sWatcher) deleteK8sServiceV1(svc *types.Service, swg *lock.StoppableWaitGroup) error {
-	k.K8sSvcCache.DeleteService(svc, swg)
-	return nil
-}
-
-func (k *K8sWatcher) addK8sEndpointV1(ep *types.Endpoints, swg *lock.StoppableWaitGroup) error {
-	k.K8sSvcCache.UpdateEndpoints(ep, swg)
-	return nil
-}
-
-func (k *K8sWatcher) updateK8sEndpointV1(oldEP, newEP *types.Endpoints, swg *lock.StoppableWaitGroup) error {
-	k.K8sSvcCache.UpdateEndpoints(newEP, swg)
-	return nil
-}
-
-func (k *K8sWatcher) deleteK8sEndpointV1(ep *types.Endpoints, swg *lock.StoppableWaitGroup) error {
-	k.K8sSvcCache.DeleteEndpoints(ep, swg)
-	return nil
 }
 
 func (k *K8sWatcher) delK8sSVCs(svc k8s.ServiceID, svcInfo *k8s.Service, se *k8s.Endpoints) error {
@@ -1424,411 +619,6 @@ func (k *K8sWatcher) addK8sSVCs(svcID k8s.ServiceID, svc *k8s.Service, endpoints
 	return nil
 }
 
-func (k *K8sWatcher) updateCiliumNetworkPolicyV2AnnotationsOnly(ciliumNPClient clientset.Interface, ciliumV2Store cache.Store, cnp *types.SlimCNP) {
-
-	scopedLog := log.WithFields(logrus.Fields{
-		logfields.CiliumNetworkPolicyName: cnp.ObjectMeta.Name,
-		logfields.K8sAPIVersion:           cnp.TypeMeta.APIVersion,
-		logfields.K8sNamespace:            cnp.ObjectMeta.Namespace,
-	})
-
-	scopedLog.Info("updating node status due to annotations-only change to CiliumNetworkPolicy")
-
-	ctrlName := cnp.GetControllerName()
-
-	// Revision will *always* be populated because importMetadataCache is guaranteed
-	// to be updated by addCiliumNetworkPolicyV2 before calls to
-	// updateCiliumNetworkPolicyV2 are invoked.
-	meta, _ := importMetadataCache.get(cnp)
-	updateContext := &k8s.CNPStatusUpdateContext{
-		CiliumNPClient:              ciliumNPClient,
-		CiliumV2Store:               ciliumV2Store,
-		NodeName:                    node.GetName(),
-		NodeManager:                 k.nodeDiscoverManager,
-		UpdateDuration:              spanstat.Start(),
-		WaitForEndpointsAtPolicyRev: k.endpointManager.WaitForEndpointsAtPolicyRev,
-	}
-
-	k8sCM.UpdateController(ctrlName,
-		controller.ControllerParams{
-			DoFunc: func(ctx context.Context) error {
-				return updateContext.UpdateStatus(ctx, cnp, meta.revision, meta.policyImportError)
-			},
-		})
-
-}
-
-func (k *K8sWatcher) addCiliumNetworkPolicyV2(ciliumNPClient clientset.Interface, ciliumV2Store cache.Store, cnp *types.SlimCNP) error {
-	scopedLog := log.WithFields(logrus.Fields{
-		logfields.CiliumNetworkPolicyName: cnp.ObjectMeta.Name,
-		logfields.K8sAPIVersion:           cnp.TypeMeta.APIVersion,
-		logfields.K8sNamespace:            cnp.ObjectMeta.Namespace,
-	})
-
-	scopedLog.Debug("Adding CiliumNetworkPolicy")
-
-	var rev uint64
-
-	rules, policyImportErr := cnp.Parse()
-	if policyImportErr == nil {
-		policyImportErr = k8s.PreprocessRules(rules, &k.K8sSvcCache)
-		// Replace all rules with the same name, namespace and
-		// resourceTypeCiliumNetworkPolicy
-		rev, policyImportErr = k.policyManager.PolicyAdd(rules, &policy.AddOptions{
-			ReplaceWithLabels: cnp.GetIdentityLabels(),
-			Source:            metrics.LabelEventSourceK8s,
-		})
-	}
-
-	if policyImportErr != nil {
-		scopedLog.WithError(policyImportErr).Warn("Unable to add CiliumNetworkPolicy")
-	} else {
-		scopedLog.Info("Imported CiliumNetworkPolicy")
-	}
-
-	// Upsert to rule revision cache outside of controller, because upsertion
-	// *must* be synchronous so that if we get an update for the CNP, the cache
-	// is populated by the time updateCiliumNetworkPolicyV2 is invoked.
-	importMetadataCache.upsert(cnp, rev, policyImportErr)
-
-	if !option.Config.DisableCNPStatusUpdates {
-		updateContext := &k8s.CNPStatusUpdateContext{
-			CiliumNPClient:              ciliumNPClient,
-			CiliumV2Store:               ciliumV2Store,
-			NodeName:                    node.GetName(),
-			NodeManager:                 k.nodeDiscoverManager,
-			UpdateDuration:              spanstat.Start(),
-			WaitForEndpointsAtPolicyRev: k.endpointManager.WaitForEndpointsAtPolicyRev,
-		}
-
-		ctrlName := cnp.GetControllerName()
-		k8sCM.UpdateController(ctrlName,
-			controller.ControllerParams{
-				DoFunc: func(ctx context.Context) error {
-					return updateContext.UpdateStatus(ctx, cnp, rev, policyImportErr)
-				},
-			},
-		)
-	}
-
-	return policyImportErr
-}
-
-func (k *K8sWatcher) deleteCiliumNetworkPolicyV2(cnp *types.SlimCNP) error {
-	scopedLog := log.WithFields(logrus.Fields{
-		logfields.CiliumNetworkPolicyName: cnp.ObjectMeta.Name,
-		logfields.K8sAPIVersion:           cnp.TypeMeta.APIVersion,
-		logfields.K8sNamespace:            cnp.ObjectMeta.Namespace,
-	})
-
-	scopedLog.Debug("Deleting CiliumNetworkPolicy")
-
-	importMetadataCache.delete(cnp)
-	ctrlName := cnp.GetControllerName()
-	err := k8sCM.RemoveControllerAndWait(ctrlName)
-	if err != nil {
-		log.Debugf("Unable to remove controller %s: %s", ctrlName, err)
-	}
-
-	_, err = k.policyManager.PolicyDelete(cnp.GetIdentityLabels())
-	if err == nil {
-		scopedLog.Info("Deleted CiliumNetworkPolicy")
-	} else {
-		scopedLog.WithError(err).Warn("Unable to delete CiliumNetworkPolicy")
-	}
-	return err
-}
-
-func (k *K8sWatcher) updateCiliumNetworkPolicyV2(ciliumNPClient clientset.Interface,
-	ciliumV2Store cache.Store,
-	oldRuleCpy, newRuleCpy *types.SlimCNP) error {
-
-	_, err := oldRuleCpy.Parse()
-	if err != nil {
-		log.WithError(err).WithField(logfields.Object, logfields.Repr(oldRuleCpy)).
-			Warn("Error parsing old CiliumNetworkPolicy rule")
-		return err
-	}
-	_, err = newRuleCpy.Parse()
-	if err != nil {
-		log.WithError(err).WithField(logfields.Object, logfields.Repr(newRuleCpy)).
-			Warn("Error parsing new CiliumNetworkPolicy rule")
-		return err
-	}
-
-	log.WithFields(logrus.Fields{
-		logfields.K8sAPIVersion:                    oldRuleCpy.TypeMeta.APIVersion,
-		logfields.CiliumNetworkPolicyName + ".old": oldRuleCpy.ObjectMeta.Name,
-		logfields.K8sNamespace + ".old":            oldRuleCpy.ObjectMeta.Namespace,
-		logfields.CiliumNetworkPolicyName:          newRuleCpy.ObjectMeta.Name,
-		logfields.K8sNamespace:                     newRuleCpy.ObjectMeta.Namespace,
-		"annotations.old":                          oldRuleCpy.ObjectMeta.Annotations,
-		"annotations":                              newRuleCpy.ObjectMeta.Annotations,
-	}).Debug("Modified CiliumNetworkPolicy")
-
-	// Do not add rule into policy repository if the spec remains unchanged.
-	if !option.Config.DisableCNPStatusUpdates {
-		if oldRuleCpy.SpecEquals(newRuleCpy.CiliumNetworkPolicy) {
-			if !oldRuleCpy.AnnotationsEquals(newRuleCpy.CiliumNetworkPolicy) {
-
-				// Update annotations within a controller so the status of the update
-				// is trackable from the list of running controllers, and so we do
-				// not block subsequent policy lifecycle operations from Kubernetes
-				// until the update is complete.
-				oldCtrlName := oldRuleCpy.GetControllerName()
-				newCtrlName := newRuleCpy.GetControllerName()
-
-				// In case the controller name changes between copies of rules,
-				// remove old controller so we do not leak goroutines.
-				if oldCtrlName != newCtrlName {
-					err := k8sCM.RemoveController(oldCtrlName)
-					if err != nil {
-						log.Debugf("Unable to remove controller %s: %s", oldCtrlName, err)
-					}
-				}
-				k.updateCiliumNetworkPolicyV2AnnotationsOnly(ciliumNPClient, ciliumV2Store, newRuleCpy)
-			}
-			return nil
-		}
-	}
-
-	return k.addCiliumNetworkPolicyV2(ciliumNPClient, ciliumV2Store, newRuleCpy)
-}
-
-func (k *K8sWatcher) updatePodHostIP(pod *types.Pod) (bool, error) {
-	if pod.SpecHostNetwork {
-		return true, fmt.Errorf("pod is using host networking")
-	}
-
-	hostIP := net.ParseIP(pod.StatusHostIP)
-	if hostIP == nil {
-		return true, fmt.Errorf("no/invalid HostIP: %s", pod.StatusHostIP)
-	}
-
-	podIP := net.ParseIP(pod.StatusPodIP)
-	if podIP == nil {
-		return true, fmt.Errorf("no/invalid PodIP: %s", pod.StatusPodIP)
-	}
-
-	hostKey := node.GetIPsecKeyIdentity()
-
-	k8sMeta := &ipcache.K8sMetadata{
-		Namespace: pod.Namespace,
-		PodName:   pod.Name,
-	}
-
-	// Initial mapping of podIP <-> hostIP <-> identity. The mapping is
-	// later updated once the allocator has determined the real identity.
-	// If the endpoint remains unmanaged, the identity remains untouched.
-	selfOwned := ipcache.IPIdentityCache.Upsert(pod.StatusPodIP, hostIP, hostKey, k8sMeta, ipcache.Identity{
-		ID:     identity.ReservedIdentityUnmanaged,
-		Source: source.Kubernetes,
-	})
-	if !selfOwned {
-		return true, fmt.Errorf("ipcache entry owned by kvstore or agent")
-	}
-
-	return false, nil
-}
-
-func (k *K8sWatcher) deletePodHostIP(pod *types.Pod) (bool, error) {
-	if pod.SpecHostNetwork {
-		return true, fmt.Errorf("pod is using host networking")
-	}
-
-	podIP := net.ParseIP(pod.StatusPodIP)
-	if podIP == nil {
-		return true, fmt.Errorf("no/invalid PodIP: %s", pod.StatusPodIP)
-	}
-
-	// a small race condition exists here as deletion could occur in
-	// parallel based on another event but it doesn't matter as the
-	// identity is going away
-	id, exists := ipcache.IPIdentityCache.LookupByIP(pod.StatusPodIP)
-	if !exists {
-		return true, fmt.Errorf("identity for IP does not exist in case")
-	}
-
-	if id.Source != source.Kubernetes {
-		return true, fmt.Errorf("ipcache entry not owned by kubernetes source")
-	}
-
-	ipcache.IPIdentityCache.Delete(pod.StatusPodIP, source.Kubernetes)
-
-	return false, nil
-}
-
-func (k *K8sWatcher) addK8sPodV1(pod *types.Pod) error {
-	logger := log.WithFields(logrus.Fields{
-		logfields.K8sPodName:   pod.ObjectMeta.Name,
-		logfields.K8sNamespace: pod.ObjectMeta.Namespace,
-		"podIP":                pod.StatusPodIP,
-		"hostIP":               pod.StatusHostIP,
-	})
-
-	skipped, err := k.updatePodHostIP(pod)
-	switch {
-	case skipped:
-		logger.WithError(err).Debug("Skipped ipcache map update on pod add")
-		return nil
-	case err != nil:
-		msg := "Unable to update ipcache map entry on pod add"
-		if err == errIPCacheOwnedByNonK8s {
-			logger.WithError(err).Debug(msg)
-		} else {
-			logger.WithError(err).Warning(msg)
-		}
-	default:
-		logger.Debug("Updated ipcache map entry on pod add")
-	}
-	return err
-}
-
-func (k *K8sWatcher) updateK8sPodV1(oldK8sPod, newK8sPod *types.Pod) error {
-	if oldK8sPod == nil || newK8sPod == nil {
-		return nil
-	}
-
-	// The pod IP can never change, it can only switch from unassigned to
-	// assigned
-	// Process IP updates
-	k.addK8sPodV1(newK8sPod)
-
-	// Check annotation updates.
-	oldAnno := oldK8sPod.GetAnnotations()
-	newAnno := newK8sPod.GetAnnotations()
-	annotationsChanged := !k8s.AnnotationsEqual([]string{annotation.ProxyVisibility}, oldAnno, newAnno)
-
-	// Check label updates too.
-	oldPodLabels := oldK8sPod.GetLabels()
-	newPodLabels := newK8sPod.GetLabels()
-	labelsChanged := !comparator.MapStringEquals(oldPodLabels, newPodLabels)
-
-	// Nothing changed.
-	if !annotationsChanged && !labelsChanged {
-		return nil
-	}
-
-	podNSName := k8sUtils.GetObjNamespaceName(&newK8sPod.ObjectMeta)
-
-	podEP := k.endpointManager.LookupPodName(podNSName)
-	if podEP == nil {
-		log.WithField("pod", podNSName).Debugf("Endpoint not found running for the given pod")
-		return nil
-	}
-
-	if labelsChanged {
-		err := updateEndpointLabels(podEP, oldPodLabels, newPodLabels)
-		if err != nil {
-			return err
-		}
-	}
-
-	if annotationsChanged {
-		podEP.UpdateVisibilityPolicy(newAnno[annotation.ProxyVisibility])
-		realizePodAnnotationUpdate(podEP)
-	}
-	return nil
-}
-
-func realizePodAnnotationUpdate(podEP *endpoint.Endpoint) {
-	// No need to log an error if the state transition didn't succeed,
-	// if it didn't succeed that means the endpoint is being deleted, or
-	// another regeneration has already been queued up for this endpoint.
-	stateTransitionSucceeded := podEP.SetState(endpoint.StateWaitingToRegenerate, "annotations updated")
-	if stateTransitionSucceeded {
-		podEP.Regenerate(&regeneration.ExternalRegenerationMetadata{
-			Reason:            "annotations updated",
-			RegenerationLevel: regeneration.RegenerateWithoutDatapath,
-		})
-	}
-}
-
-func updateEndpointLabels(ep *endpoint.Endpoint, oldLbls, newLbls map[string]string) error {
-	newLabels := labels.Map2Labels(newLbls, labels.LabelSourceK8s)
-	newIdtyLabels, _ := labels.FilterLabels(newLabels)
-	oldLabels := labels.Map2Labels(oldLbls, labels.LabelSourceK8s)
-	oldIdtyLabels, _ := labels.FilterLabels(oldLabels)
-
-	err := ep.ModifyIdentityLabels(newIdtyLabels, oldIdtyLabels)
-	if err != nil {
-		log.WithError(err).Debugf("error while updating endpoint with new labels")
-		return err
-	}
-
-	log.WithFields(logrus.Fields{
-		logfields.EndpointID: ep.GetID(),
-		logfields.Labels:     logfields.Repr(newIdtyLabels),
-	}).Debug("Updated endpoint with new labels")
-	return nil
-
-}
-
-func (k *K8sWatcher) deleteK8sPodV1(pod *types.Pod) error {
-	logger := log.WithFields(logrus.Fields{
-		logfields.K8sPodName:   pod.ObjectMeta.Name,
-		logfields.K8sNamespace: pod.ObjectMeta.Namespace,
-		"podIP":                pod.StatusPodIP,
-		"hostIP":               pod.StatusHostIP,
-	})
-
-	skipped, err := k.deletePodHostIP(pod)
-	switch {
-	case skipped:
-		logger.WithError(err).Debug("Skipped ipcache map delete on pod delete")
-	case err != nil:
-		logger.WithError(err).Warning("Unable to delete ipcache map entry on pod delete")
-	default:
-		logger.Debug("Deleted ipcache map entry on pod delete")
-	}
-	return err
-}
-
-func (k *K8sWatcher) updateK8sV1Namespace(oldNS, newNS *types.Namespace) error {
-	if oldNS == nil || newNS == nil {
-		return nil
-	}
-
-	// We only care about label updates
-	if comparator.MapStringEquals(oldNS.GetLabels(), newNS.GetLabels()) {
-		return nil
-	}
-
-	oldNSLabels := map[string]string{}
-	newNSLabels := map[string]string{}
-
-	for k, v := range oldNS.GetLabels() {
-		oldNSLabels[policy.JoinPath(ciliumio.PodNamespaceMetaLabels, k)] = v
-	}
-	for k, v := range newNS.GetLabels() {
-		newNSLabels[policy.JoinPath(ciliumio.PodNamespaceMetaLabels, k)] = v
-	}
-
-	oldLabels := labels.Map2Labels(oldNSLabels, labels.LabelSourceK8s)
-	newLabels := labels.Map2Labels(newNSLabels, labels.LabelSourceK8s)
-
-	oldIdtyLabels, _ := labels.FilterLabels(oldLabels)
-	newIdtyLabels, _ := labels.FilterLabels(newLabels)
-
-	eps := k.endpointManager.GetEndpoints()
-	failed := false
-	for _, ep := range eps {
-		epNS := ep.GetK8sNamespace()
-		if oldNS.Name == epNS {
-			err := ep.ModifyIdentityLabels(newIdtyLabels, oldIdtyLabels)
-			if err != nil {
-				log.WithError(err).WithField(logfields.EndpointID, ep.ID).
-					Warningf("unable to update endpoint with new namespace labels")
-				failed = true
-			}
-		}
-	}
-	if failed {
-		return errors.New("unable to update some endpoints with new namespace labels")
-	}
-	return nil
-}
-
 // K8sEventProcessed is called to do metrics accounting for each processed
 // Kubernetes event
 func (k *K8sWatcher) K8sEventProcessed(scope string, action string, status bool) {
@@ -1846,64 +636,4 @@ func (k *K8sWatcher) K8sEventReceived(scope string, action string, valid, equal 
 	k8smetrics.LastInteraction.Reset()
 
 	metrics.KubernetesEventReceived.WithLabelValues(scope, action, strconv.FormatBool(valid), strconv.FormatBool(equal)).Inc()
-}
-
-func endpointUpdated(endpoint *types.CiliumEndpoint) {
-	// default to the standard key
-	encryptionKey := node.GetIPsecKeyIdentity()
-
-	id := identity.ReservedIdentityUnmanaged
-	if endpoint.Identity != nil {
-		id = identity.NumericIdentity(endpoint.Identity.ID)
-	}
-
-	if endpoint.Encryption != nil {
-		encryptionKey = uint8(endpoint.Encryption.Key)
-	}
-
-	if endpoint.Networking != nil {
-		if endpoint.Networking.NodeIP == "" {
-			// When upgrading from an older version, the nodeIP may
-			// not be available yet in the CiliumEndpoint and we
-			// have to wait for it to be propagated
-			return
-		}
-
-		nodeIP := net.ParseIP(endpoint.Networking.NodeIP)
-		if nodeIP == nil {
-			log.WithField("nodeIP", endpoint.Networking.NodeIP).Warning("Unable to parse node IP while processing CiliumEndpoint update")
-			return
-		}
-
-		k8sMeta := &ipcache.K8sMetadata{
-			Namespace: endpoint.Namespace,
-			PodName:   endpoint.Name,
-		}
-
-		for _, pair := range endpoint.Networking.Addressing {
-			if pair.IPV4 != "" {
-				ipcache.IPIdentityCache.Upsert(pair.IPV4, nodeIP, encryptionKey, k8sMeta,
-					ipcache.Identity{ID: id, Source: source.CustomResource})
-			}
-
-			if pair.IPV6 != "" {
-				ipcache.IPIdentityCache.Upsert(pair.IPV6, nodeIP, encryptionKey, k8sMeta,
-					ipcache.Identity{ID: id, Source: source.CustomResource})
-			}
-		}
-	}
-}
-
-func endpointDeleted(endpoint *types.CiliumEndpoint) {
-	if endpoint.Networking != nil {
-		for _, pair := range endpoint.Networking.Addressing {
-			if pair.IPV4 != "" {
-				ipcache.IPIdentityCache.Delete(pair.IPV4, source.CustomResource)
-			}
-
-			if pair.IPV6 != "" {
-				ipcache.IPIdentityCache.Delete(pair.IPV6, source.CustomResource)
-			}
-		}
-	}
 }
